% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_score}
\alias{compute_score}
\title{Compute an Individual Scoring Metric}
\usage{
compute_score(metric, preds, truth, alpha = 0.05)
}
\description{
Compute an Individual Scoring Metric
}
\keyword{internal}
